# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LibTrack System API",
        "description": "API documentation for the LibTrack Library Tracking System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/admin/borrowings": {
            "get": {
                "tags": [
                    "Borrowing Management"
                ],
                "summary": "Get all borrowing records",
                "description": "Get all borrowing records.",
                "operationId": "6988a4fcbb4200fc5232f7679ecb77c8",
                "responses": {
                    "200": {
                        "description": "Borrowing records retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Borrowing Management"
                ],
                "summary": "Borrow a book",
                "description": "Borrow a book.",
                "operationId": "f14526da76d72ab46d8d860cad93c4f8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "book_id",
                                    "student_id",
                                    "due_date"
                                ],
                                "properties": {
                                    "book_id": {
                                        "type": "string",
                                        "example": "BOOK_MONGODB_ID"
                                    },
                                    "student_id": {
                                        "type": "string",
                                        "example": "STUDENT_MONGODB_ID"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-09-20"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book borrowed successfully"
                    },
                    "400": {
                        "description": "Book unavailable"
                    },
                    "404": {
                        "description": "Book or student not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/borrowings/active": {
            "get": {
                "tags": [
                    "Borrowing Management"
                ],
                "summary": "Get active borrowings",
                "description": "Get active borrowings.",
                "operationId": "1e29f99675f3b0a3185984415d85dacd",
                "responses": {
                    "200": {
                        "description": "Active borrowings retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/borrowings/{id}": {
            "get": {
                "tags": [
                    "Borrowing Management"
                ],
                "summary": "Get a single borrowing record",
                "description": "Get a single borrowing record.",
                "operationId": "6c3d68dc120535f44ac186ed10709ee3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Borrowing record retrieved successfully"
                    },
                    "404": {
                        "description": "Borrowing record not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/borrowings/{id}/return": {
            "put": {
                "tags": [
                    "Borrowing Management"
                ],
                "summary": "Return a borrowed book",
                "description": "Return a borrowed book.",
                "operationId": "7f1fd1e4b24cce893fc301553903defd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book returned successfully"
                    },
                    "400": {
                        "description": "Book already returned"
                    },
                    "404": {
                        "description": "Borrowing record not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Bearer Token",
                "scheme": "bearer"
            }
        }
    },
...
}